requirements.txt
fastapi>=0.135.0
uvicorn>=0.32.0
sqlalchemy>=2.0.0
apscheduler>=3.10.0
anthropic>=0.40.0
httpx>=0.27.0
python-dotenv>=1.0.0
curl_cffi>=0.7.0
